requirements.txt
numpy==1.26.4
scipy==1.15.3
scikit-learn==1.7.2
PyYAML==6.0.3
Pillow==12.1.0
timm==1.0.21
easydict==1.13
h5py==3.15.1
ftfy==6.3.1
regex==2025.11.3
termcolor==3.2.0
matplotlib==3.10.8
tqdm==4.67.1
faiss-cpu==1.13.0
open3d==0.19.0
rbo==0.1.3
